requirements.txt
psycopg2-binary==2.7.*
celery==4.1.*
requests==2.18.*
iso-639==0.4.*
jsonschema==2.6.*
pillow==5.0.*
PyYaml==3.12.*
psutil==5.4.*

coverage
coverage-badge
cairosvg

Django==2.0.*

djangorestframework==3.8.*
django-rest-swagger==2.1.*
django-debug-toolbar==1.9.*
django-redis==4.9.*
django-allauth==0.35.*
django-positions==0.5.*